+++ file-formats/graphics/wow-blp/CHANGELOG.md
@@ -5,11 +5,11 @@ All notable changes to wow-blp will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.1] - 2025-01-28
+## [0.3.0] - 2025-08-07
 
 ### Changed
 
-- Version bump to 0.2.1 for coordinated workspace release
+- Version bump to 0.3.0 for coordinated workspace release
 - Updated dependencies and documentation
 
 ## [0.2.0] - 2025-06-28